requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
django-filter==22.1
validate-docbr==1.10.0
gunicorn==20.0.4
pymysql==1.0.2